refactor [server]: create common module
