+++ server/src/app.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 
+import { CommonModule } from './common/common.module';
 import { TypeormModule } from './infra/database/typeorm.module';
 import { ArticleModule } from './modules/article/article.module';
 import { AuthModule } from './modules/auth/auth.module';
-import { UserModule } from './modules/user/user.module';
 import { StatusModule } from './modules/status/status.module';
+import { UserModule } from './modules/user/user.module';
 
 @Module({
     imports: [
@@ -15,6 +16,7 @@ import { StatusModule } from './modules/status/status.module';
         UserModule,
         ArticleModule,
         StatusModule,
+        CommonModule,
     ],
     controllers: [],
     providers: [],
