--- server/src/modules/auth/guard/local-auth.guard.ts
@@ -1,6 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
-
-// Runs local strategy before controller method
-@Injectable()
-export class LocalAuthGuard extends AuthGuard('local') {}
